backend/requirements.txt
fastapi
uvicorn
httpx
openai
anthropic
google-generativeai
python-dotenv
pydantic
